catching errors in gcloud app

+++ gcloud-app/app.py
@@ -21,7 +21,10 @@ def validate_post():
   data = request.json
   print(data)
   if 'REMOVEUSER' in data:
-    return deleteSecrets(data)
+    if 'REMOVEUSER' in data == '*':
+      return resetApp()
+    else:
+      return deleteSecrets(data)
   elif 'NEWUSER' in data:
     return updateSecrets(data)
   else:
@@ -88,7 +91,30 @@ def deleteSecrets(data):
     finally:
         return print("User Deleted")
 
+def resetApp():
+    projectId = os.environ['PROJECT_ID']
+    access_token = get_access_token()
+    gcloud_path = shutil.which("gcloud")
+    setSecrets = "--set-secrets=AWS_ACCESS_KEY_ID=projects/$PROJECT_NUMBER/secrets/derf-default-accessKeyId-AWS:latest,AWS_SECRET_ACCESS_KEY=projects/$PROJECT_NUMBER/secrets/derf-default-accessKeySecret-AWS:latest,AWS_ACCESS_KEY_ID_USER02=projects/$PROJECT_NUMBER/secrets/derf-user02-accessKeyId-AWS:latest,AWS_SECRET_ACCESS_KEY_USER02=projects/$PROJECT_NUMBER/secrets/derf-user02-accessKeySecret-AWS:latest,AWS_ACCESS_KEY_ID_USER01=projects/$PROJECT_NUMBER/secrets/derf-user01-accessKeyId-AWS:latest,AWS_SECRET_ACCESS_KEY_USER01=projects/$PROJECT_NUMBER/secrets/derf-user01-accessKeySecret-AWS:latest"
 
+    try:
+        completedProcess = subprocess.run("$GCLOUD run services update aws-proxy-app $SETSECRETS --region us-central1 --project $PROJECT_ID --access-token-file $CLOUDSDK_AUTH_ACCESS_TOKEN", 
+                                          env={"GCLOUD": gcloud_path, "SETSECRETS": setSecrets, "PROJECT_ID": projectId, "CLOUDSDK_AUTH_ACCESS_TOKEN": access_token},
+                                          shell=True, 
+                                          timeout=180,
+                                          text=True
+                                          )
+
+        response = print("Proxy app redeployed and reset to original configuration")
+        return response
+    except subprocess.CalledProcessError as e:
+        response = print("Process error when deleting users")
+        return response
+    except subprocess.TimeoutExpired as e:
+        response = print("Unable to delete user, process timedout")
+        return response
+    finally:
+        return print("User Deleted")
 
 def get_access_token():
